# openapi.yaml
openapi: 3.0.2
info:
  description: Moov ACH ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing, and validating ACH files. ACH is the primary method of electronic money movement throughout the United States.
  version: v1
  title: ACH API
  contact:
    url: https://github.com/moov-io/ach
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /files:
    get:
      tags: ['ACH Files']
      summary: List Files
      description: List all ACH Files created with the ACH service. These Files are not persisted through multiple runs of the service.
      operationId: getFiles
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
      responses:
        '200':
          description: A list of File objects
          headers:
            X-Total-Count:
              description: The total number of Originators
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Files'
  /files/{fileID}:
    post:
      tags: ['ACH Files']
      summary: Create File
      description: Create a new File object from either the plaintext or JSON representation.
      operationId: createFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
        - name: skipAll
          in: query
          description: Optional parameter to disable all validation checks for a File
          schema:
            type: boolean
        - name: requireABAOrigin
          in: query
          description: Optional parameter to configure ImmediateOrigin validation
          schema:
            type: boolean
        - name: bypassOrigin
          in: query
          description: Optional parameter to configure ImmediateOrigin validation
          schema:
            type: boolean
          deprecated: true
        - name: bypassOriginValidation
          in: query
          description: Optional parameter to configure ImmediateOrigin validation
          schema:
            type: boolean
        - name: bypassDestination
          in: query
          description: Optional parameter to configure ImmediateDestination validation
          schema:
            type: boolean
          deprecated: true
        - name: bypassDestinationValidation
          in: query
          description: Optional parameter to configure ImmediateDestination validation
          schema:
            type: boolean
        - name: customTraceNumbers
          in: query
          description: Optional parameter to configure ImmediateDestination validation
          schema:
            type: boolean
        - name: allowZeroBatches
          in: query
          description: Optional parameter to configure ImmediateDestination validation
          schema:
            type: boolean
        - name: allowMissingFileHeader
          in: query
          description: Optional parameter to configure ImmediateDestination validation
          schema:
            type: boolean
        - name: allowMissingFileControl
          in: query
          description: Optional parameter to configure ImmediateDestination validation
          schema:
            type: boolean
        - name: bypassCompanyIdentificationMatch
          in: query
          description: Optional parameter to configure ImmediateDestination validation
          schema:
            type: boolean
        - name: customReturnCodes
          in: query
          description: Optional parameter to configure ImmediateDestination validation
          schema:
            type: boolean
        - name: unequalServiceClassCode
          in: query
          description: Optional parameter to configure ImmediateDestination validation
          schema:
            type: boolean
        - name: unorderedBatchNumbers
          in: query
          description: Allow a file to be read with unordered batch numbers.
          schema:
            type: boolean
          deprecated: true
        - name: allowUnorderedBatchNumbers
          in: query
          description: Allow a file to be read with unordered batch numbers.
          schema:
            type: boolean
        - name: allowInvalidCheckDigit
          in: query
          description: Allow the CheckDigit field in EntryDetail to differ from the expected calculation
          schema:
            type: boolean
        - name: unequalAddendaCounts
          in: query
          description: Optional parameter to configure UnequalAddendaCounts validation
          schema:
            type: boolean
        - name: preserveSpaces
          in: query
          description: Optional parameter to save all padding spaces
          schema:
            type: boolean
        - name: allowInvalidAmounts
          in: query
          description: Optional parameter to save all padding spaces
          schema:
            type: boolean
        - name: allowSpecialCharacters
          in: query
          description: Optional parameter to permit a wider range of UTF-8 characters in alphanumeric fields
          schema:
            type: boolean
      requestBody:
        description: Content of the ACH file (in json or raw text)
        required: true
        content:
          text/plain:
            schema:
              description: A plaintext ACH file
              type: string
              example: |
               101 23138010401210428821906240000A094101Federal Reserve Bank   My Bank Name
               5225Name on Account                     121042882 PPDREG.SALARY      190625   1121042880000001
               62723138010412345678         0100000000               Receiver Account Name   0121042880000001
               82250000010023138010000100000000000000000000121042882                          121042880000001
               9000001000001000000010023138010000100000000000000000000
               9999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
               9999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
               9999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
               9999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
               9999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFile'
      responses:
        '200':
          description: A JSON object containing a new File
          headers:
            Location:
              description: The location of the new resource
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateFileResponse'
        '400':
          description: "Invalid File Header Object"
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
    get:
      tags: ['ACH Files']
      summary: Retrieve File
      description: Get the details of an existing File using the unique File identifier that was returned upon creation.
      operationId: getFileByID
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: A File object for the supplied ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '404':
          description: A resource with the specified ID was not found
    delete:
      tags: ['ACH Files']
      summary: Delete File
      description: Permanently deletes a File and associated Batches. It cannot be undone.
      operationId: deleteACHFile
      parameters:
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
      responses:
        '200':
          description: Permanently deleted File.
        '404':
          description: A File with the specified ID was not found.
  /files/{fileID}/build:
    get:
      tags: ['ACH Files']
      summary: Build File
      description: |
        Assembles the existing File (batches and controls) records, computes sequence numbers and totals. Returns JSON formatted file.
      operationId: buildFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: File built successfully without errors.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/File'
  /files/{fileID}/contents:
    get:
      tags: ['ACH Files']
      summary: Get File Contents
      description: |
        Assembles the existing File (batches and controls) records, computes sequence numbers and totals. Returns plaintext file.
      operationId: getFileContents
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
        - name: X-Line-Ending
          in: header
          description: Optional Line Ending allows application developer to alter line ending of the file data returned. Supported choices are LF (Line Feed) and CRLF (Carriage Return Line Feed).
          example: "CRLF"
          schema:
            type: string
            enum: ["LF", "CRLF"]
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: File built successfully without errors.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RawFile'
  /files/{fileID}/validate:
    parameters:
      - name: skipAll
        in: query
        description: Optional parameter to disable all validation checks for a File
        schema:
          type: boolean
      - name: requireABAOrigin
        in: query
        description: Optional parameter to configure ImmediateOrigin validation
        schema:
          type: boolean
      - name: bypassOriginValidation
        in: query
        description: Optional parameter to configure ImmediateOrigin validation
        schema:
          type: boolean
      - name: bypassDestinationValidation
        in: query
        description: Optional parameter to configure ImmediateDestination validation
        schema:
          type: boolean
      - name: customTraceNumbers
        in: query
        description: Optional parameter to configure ImmediateDestination validation
        schema:
          type: boolean
      - name: allowZeroBatches
        in: query
        description: Optional parameter to configure ImmediateDestination validation
        schema:
          type: boolean
      - name: allowMissingFileHeader
        in: query
        description: Optional parameter to configure ImmediateDestination validation
        schema:
          type: boolean
      - name: allowMissingFileControl
        in: query
        description: Optional parameter to configure ImmediateDestination validation
        schema:
          type: boolean
      - name: bypassCompanyIdentificationMatch
        in: query
        description: Optional parameter to configure ImmediateDestination validation
        schema:
          type: boolean
      - name: customReturnCodes
        in: query
        description: Optional parameter to configure ImmediateDestination validation
        schema:
          type: boolean
      - name: unequalServiceClassCode
        in: query
        description: Optional parameter to configure ImmediateDestination validation
        schema:
          type: boolean
      - name: allowUnorderedBatchNumbers
        in: query
        description: Allow a file to be read with unordered batch numbers.
        schema:
          type: boolean
      - name: allowInvalidCheckDigit
        in: query
        description: Allow the CheckDigit field in EntryDetail to differ from the expected calculation
        schema:
          type: boolean
      - name: unequalAddendaCounts
        in: query
        description: Optional parameter to configure UnequalAddendaCounts validation
        schema:
          type: boolean
      - name: preserveSpaces
        in: query
        description: Optional parameter to save all padding spaces
        schema:
          type: boolean
      - name: allowInvalidAmounts
        in: query
        description: Optional parameter to save all padding spaces
        schema:
          type: boolean
      - name: allowSpecialCharacters
        in: query
        description: Optional parameter to permit a wider range of UTF-8 characters in alphanumeric fields
        schema:
          type: boolean
    get:
      tags: ['ACH Files']
      summary: Validate File
      description: Validates the existing File. You need only supply the unique File identifier that was returned upon creation.
      operationId: checkFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: File validated successfully without errors.
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
        '400':
          description: Validation failed. Check response for errors
    post:
      tags: ['ACH Files']
      summary: Validate File (Custom)
      description: Validates the existing File. You need only supply the unique File identifier that was returned upon creation.
      operationId: validateFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateOpts'
      responses:
        '200':
          description: File validated successfully without errors.
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
        '400':
          description: Validation failed. Check response for errors
  /files/{fileID}/segment:
    post:
      tags: ['ACH Files']
      summary: Segment FileID
      description: Split one FileID into two. One with only debits and one with only credits.
      operationId: segmentFileID
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      requestBody:
        description: Optional configuration for segmenting files
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SegmentFileConfiguration'
      responses:
        '200':
          description: An ID of each new ACH file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SegmentedFiles'
        '400':
          description: See error in response body
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
        '404':
          description: A resource with the specified ID was not found
  /files/{fileID}/flatten:
    post:
      tags: ['ACH Files']
      summary: Flatten Batches
      description: Consolidate Batches and Entries into the minimum number of Batches needed.
      operationId: flattenFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: An ID of the new ACH file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlattenFileResponse'
        '400':
          description: See error in response body
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
        '404':
          description: A resource with the specified ID was not found
  /files/{fileID}/batches:
    get:
      tags: ['ACH Files']
      summary: Get Batches
      description: Get the Batches on a File.
      operationId: getFileBatches
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      responses:
        '200':
          description: A list of Batch objects
          headers:
            X-Total-Count:
              description: The total number of Batches on the File.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batches'
    post:
      tags: ['ACH Files']
      summary: Append Batch to File
      description: Append a Batch record to the specified File.
      operationId: addBatchToFile
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: "3f2d23ee214"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Batch'
      responses:
        '200':
          description: Batch added to File
        '400':
          description: See error in response body
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
        '404':
          description: A resource with the specified ID was not found
...
  /merge:
    post:
      tags: ['ACH Files']
      summary: Merge Files
      description: Combine multiple fileIDs and files together. Returned is the results of merging
      operationId: mergeFiles
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the system's logs
          example: "rs4f9915"
          schema:
            type: string
      requestBody:
        description: JSON object containing FileIDs and ACH files (in JSON format)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeFilesRequest'
      responses:
        '200':
          description: Files were merged together
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MergeFilesResponse'
        '400':
          description: See error in response body
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
        '404':
          description: A resource with the specified ID was not found
...
components:
  schemas:
    CreateFile:
      description: \`batches\` OR `IATBatches` is required
      properties:
        id:
          type: string
          description: File ID
          example: "3f2d23ee214"
        fileHeader:
          $ref: '#/components/schemas/FileHeader'
        batches:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
          minItems: 1
        IATBatches:
          type: array
          items:
            $ref: '#/components/schemas/IATBatch'
          minItems: 1
        fileControl:
          $ref: '#/components/schemas/FileControl'
        advEntryDetails:
          type: array
          items:
            $ref: '#/components/schemas/ADVEntryDetail'
        advBatchControl:
          $ref: '#/components/schemas/ADVBatchControl'
      required:
        - fileHeader
        - batches
        - IATBatches
    CreateFileResponse:
      properties:
        id:
          type: string
          description: File ID
          example: "1e522dc8"
        file:
          $ref: '#/components/schemas/File'
        error:
          type: string
          description: An error message describing the problem intended for humans.
          example: Validation error(s) present.
...
    Files:
      type: array
      items:
        $ref: '#/components/schemas/File'
...
    SegmentFileConfiguration:
      properties: {} # TODO: Are there any config options people need?
...